% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_Rbase.R
\name{lm_Rbase}
\alias{lm_Rbase}
\title{Calcola coef modello lineare}
\usage{
lm_Rbase(formula, data)
}
\arguments{
\item{formula}{un oggetto della classe formula}

\item{data}{dataset contenente le variabili}
}
\value{
\code{vector}
}
\description{
Calcola la soluzione OLS di un modello lineare con funzione di perdita data
dalla somma dei residui al quadrato
}
