% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_rcpp_arma.R
\name{lm_rcpp_arma}
\alias{lm_rcpp_arma}
\title{Calcola coef modello lineare con RcppArmadillo}
\usage{
lm_rcpp_arma(formula, data)
}
\arguments{
\item{formula}{un oggetto della classe formula}

\item{data}{dataset contenente le variabili}
}
\value{
\code{vector}
}
\description{
Calcola la soluzione OLS di un modello lineare con funzione di perdita data
dalla somma dei residui al quadrato
}
