% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_rcpp_arma}
\alias{rw_rcpp_arma}
\title{Passeggiata casuale con RcppArmadillo}
\usage{
rw_rcpp_arma(t)
}
\arguments{
\item{t}{numero di simulazioni}
}
\value{
\code{int}
}
\description{
Simula il valore al tempo t di una passeggiata casuale
}
